[Refactor] Refactor folder & files structure, add cubit state management
