+++ first_flutter_app/lib/app.dart
@@ -1,0 +1,23 @@
+import 'package:first_flutter_app/modules/dashboard/cubit/app_cubit.dart';
+import 'package:first_flutter_app/modules/dashboard/screens/home_screen/home_screen.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => AppCubit(),
+      child: MaterialApp(
+        title: 'First Flutter App',
+        theme: ThemeData(
+          useMaterial3: true,
+          colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepOrange),
+        ),
+        home: MyHomePage(),
+      ),
+    );
+  }
+}
