+++ first_flutter_app/lib/modules/dashboard/screens/home_screen/favorite_screen/favorite_screen.dart
@@ -1,0 +1,37 @@
+import 'package:first_flutter_app/modules/dashboard/cubit/app_cubit.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class FavoritesScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final appState = context.watch<AppCubit>().state;
+
+    if (appState.favorites.isEmpty) {
+      return Center(
+        child: Text('No favorites yet.'),
+      );
+    }
+
+    return ListView(
+      children: [
+        Padding(
+          padding: const EdgeInsets.all(20),
+          child: Text('You have '
+              '${appState.favorites.length} favorites:'),
+        ),
+        for (var pair in appState.favorites)
+          ListTile(
+            leading: Icon(Icons.favorite),
+            title: Text(pair.asLowerCase),
+            trailing: IconButton(
+              icon: Icon(Icons.delete),
+              onPressed: () {
+                context.read<AppCubit>().removeFavorite(pair);
+              },
+            ),
+          ),
+      ],
+    );
+  }
+}
