+++ first_flutter_app/pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
+  flutter_bloc: ^8.1.6
 
 dev_dependencies:
   flutter_test:
